# sql: read backward-comptatible data using colfamilies