@RaduBerinde 2016-06-15
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 2 files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[sql/sqlbase/rowfetcher.go, line 329 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKAjmjYnnIZlFWyrSr:-KKKAjmjYnnIZlFWyrSs:627436435) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L329)):*
> ```Go
> }
> 
> func (rf *RowFetcher) processValueSingle(
> ```

This needs a comment, especially explaining that we populate `rf.row`

---

*[sql/sqlbase/rowfetcher.go, line 345 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKBR7LxlQsBCV-jvv6:-KKKBR7MI_ODeI4R6EaC:2001342991) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L345)):*
> ```Go
> 		}
> 		kind := rf.cols[idx].Type.Kind
> 		value, err := UnmarshalColumnValue(&rf.alloc, kind, kv.Value)
> ```

NYC but seems odd that we have UnmarshalColumnValue and DecodeTableValue which should be doing pretty much the same thing.

---

*[sql/sqlbase/rowfetcher.go, line 370 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKAgeOXv7geor8-kN7:-KKKAgeOXv7geor8-kN8:627436435) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L370)):*
> ```Go
> }
> 
> func (rf *RowFetcher) processValueTuple(
> ```

This needs a comment, especially explaining that we populate `rf.row`

---

*[sql/sqlbase/rowfetcher.go, line 396 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKALLJvgDeeU5LcRqs:-KKKALLJvgDeeU5LcRqt:1540374240) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L396)):*
> ```Go
> 		if !ok || !rf.valNeededForCol[idx] {
> 			// This column wasn't requested, so read its length and skip it.
> 			i, err := roachpb.PeekValueLength(tupleBytes)
> ```

Ideally rowFetcher would generate `EncDatum`s instead of `Datum`s and that would make the logic simpler (we won't need `valNeededForCol` at all, it would be up to the user of the class to decide if they want to decode them or not.   We will need to do that at some point, hopefully it won't cause any regressions.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-:-KKKC0TEvWRb2rYqKflC:1417088721)*
<!-- Sent from Reviewable.io -->