@paperstreet 2016-06-15
TFTR!

---

Review status: 0 of 2 files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[sql/sqlbase/rowfetcher.go, line 329 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKAjmjYnnIZlFWyrSr:-KKKEITyBP-tYT9fq8Q9:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L329)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> This needs a comment, especially explaining that we populate `rf.row`
</details>
Done.

---

*[sql/sqlbase/rowfetcher.go, line 345 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKBR7LxlQsBCV-jvv6:-KKKGUSyBW8HaGMyMjqy:385190443) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L345)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> NYC but seems odd that we have UnmarshalColumnValue and DecodeTableValue which should be doing pretty much the same thing.
</details>
Great eye. I think this can go away if we don't change the tuple encoding, but I think that will end up happening, so keeping this for now. TODOed

---

*[sql/sqlbase/rowfetcher.go, line 370 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKAgeOXv7geor8-kN7:-KKKF3RKCyD8rzNPy_Av:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L370)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> This needs a comment, especially explaining that we populate `rf.row`
</details>
Done.

---

*[sql/sqlbase/rowfetcher.go, line 396 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-KKKALLJvgDeeU5LcRqs:-KKKFIyqOQmdA-xPOdbU:2133586090) ([raw file](https://github.com/cockroachdb/cockroach/blob/9a8cc155375c0a8b1f9adf73b8ceb95cb8a2989b/sql/sqlbase/rowfetcher.go#L396)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> Ideally rowFetcher would generate `EncDatum`s instead of `Datum`s and that would make the logic simpler (we won't need `valNeededForCol` at all, it would be up to the user of the class to decide if they want to decode them or not.   We will need to do that at some point, hopefully it won't cause any regressions.
</details>
TODOed

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7243#-:-KKKOZN2cRtHrXwvj5dA:-485616692)*
<!-- Sent from Reviewable.io -->